Align CLI JSON tests with repo typing and import conventions
- move concurrency-limit ls tests into tests/cli/test_global_concurrency_limit.py

- remove standalone tests/cli/test_concurrency_limit.py

- avoid deferred json imports in updated test modules
